refactor: create narrowing functions
