+++ src/narrowing/index.ts
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2021, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+export {
+  isObject,
+  isPrimitive,
+  isArray,
+  isNull,
+  isBoolean,
+  isEmpty,
+  isRow,
+  isColumn,
+  isNumber,
+  isString,
+  isTableConfig,
+  isTitle,
+  isTestResult,
+  isNotQuotable,
+  isValidTestRunID,
+  isValidApexClassID
+} from './narrowing';
