+++ src/tests/utils.ts
@@ -6,23 +6,8 @@
  */
 
 import { Connection } from '@salesforce/core';
-import { CLASS_ID_PREFIX, TEST_RUN_ID_PREFIX } from './constants';
 import { NamespaceInfo } from './types';
 import { QueryResult } from 'jsforce';
-
-export function isValidTestRunID(testRunId: string): boolean {
-  return (
-    (testRunId.length === 15 || testRunId.length === 18) &&
-    testRunId.startsWith(TEST_RUN_ID_PREFIX)
-  );
-}
-
-export function isValidApexClassID(apexClassId: string): boolean {
-  return (
-    (apexClassId.length === 15 || apexClassId.length === 18) &&
-    apexClassId.startsWith(CLASS_ID_PREFIX)
-  );
-}
 
 export function calculatePercentage(dividend: number, divisor: number): string {
   let percentage = '0%';
