+++ src/utils/types.ts
@@ -66,3 +66,20 @@ export type DataRecordResult = {
 export type IdRecord = {
   Id: string;
 };
+
+export interface Row {
+  [column: string]: string;
+}
+
+export interface Column {
+  key: string;
+  label: string;
+}
+
+export interface Title {
+  title: string;
+}
+
+export type TableConfig = Title & {
+  columns: Column[];
+};
